% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sym.R
\name{add_sym}
\alias{add_sym}
\title{add symbols in rowData to a SummarizedExperiment that has Entrez IDs for rownames}
\usage{
add_sym(x)
}
\arguments{
\item{x}{SummarizedExperiment instance}
}
\description{
add symbols in rowData to a SummarizedExperiment that has Entrez IDs for rownames
}
\note{
Will fail if 'symbol' is a column of rowData(x)
}
\examples{
if (interactive()) {
  bq = pancan_BQ()
  rnse = try(buildPancanSE(bq, assay="RNASeqv2"))
  if (inherits(rnse, "try-error")) stop("probably need CGC_BILLING set in environment or with pancan_BQ")
  add_sym(rnse)
}
}
