% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotTabs}
\alias{annotTabs}
\title{table names in Annotated pancancer data release}
\format{character vector}
\source{
pancancer-atlas in BigQuery
}
\usage{
annotTabs
}
\description{
table names in Annotated pancancer data release
}
\examples{
BiocOncoTK::annotTabs
}
\keyword{datasets}
