% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipg_tests.R
\name{bipg_tests}
\alias{bipg_tests}
\title{configure a bipartite graph relating tumor type to gene, using graphNEL}
\usage{
bipg_tests(stattab, genes_adverse = NA, genes_favorable = NA,
  gpar_cex = 0.65, gpar_lwd = 0)
}
\arguments{
\item{stattab}{a data.frame with columns 'tumor', 'gene', and 'tstat'}

\item{genes_adverse}{a vector of genes whose increased expression is 
regarded as adverse}

\item{genes_favorable}{a vector of genes whose increased expression is 
regarded as favorable}

\item{gpar_cex}{tune size of graph labels}

\item{gpar_lwd}{tune appearance of node boundaries}
}
\value{
a graphNEL instance (graph package)
}
\description{
configure a bipartite graph relating tumor type to gene, using graphNEL
}
\examples{
bipg_tests(k23sig)
}
