% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbuilder.R
\name{buildPancanSE}
\alias{buildPancanSE}
\title{helper for SummarizedExperiment construction from pancan}
\usage{
buildPancanSE(bq, acronym = "BLCA", assay = "meth450k",
  sampType = "TP", subjectIDName = "ParticipantBarcode",
  seTransform = force, bindMethRowranges = TRUE,
  featIDMap = featIDMapper())
}
\arguments{
\item{bq}{instance of BigQueryConnection for pancancer-atlas.Annotated Dataset}

\item{acronym}{character(1) 'cohort' label, e.g., 'BLCA'}

\item{assay}{character(1) element from names(BiocOncoTK::annotTabs), e.g., 'meth450k'.
If `assay == "mc3_MAF"` an error is thrown as the mutation data are
inconsistently annotated; the message produced directs the user to
`mc3toGR`.}

\item{sampType}{character(1) element from 
BiocOncoTK::pancan_sampTypeMap$"SampleTypeLetterCode", 
e.g., 'TP' for Primary solid Tumor samples,
or 'TB' for peripheral blood sample from primary blood derived cancer}

\item{subjectIDName}{character(1) field name for subject identifier}

\item{seTransform}{a function that accepts a SummarizedExperiment and returns a SummarizedExperiment; useful for feature name remapping, defaults to force (does nothing)}

\item{bindMethRowranges}{logical(1) if true and assay is meth27k}

\item{featIDMap}{a named character() vector defining, for each
assay type, what field should be used to label features in rownames.
or meth450k, annotation from FDb.InfiniumMethylation.hg19
and EnsDb.Hsapiens.v75 is obtained for available features
and bound into the rowRanges component of returned object}
}
\value{
SummarizedExperiment, with metadata on acronym, assay,
and sampleType propagated; if the assay is a methylation
assay and bindMethRowranges is TRUE, a RangedSummarizedExperiment
is returned.
}
\description{
helper for SummarizedExperiment construction from pancan
}
\note{
Note that pancancer-atlas is distinguished from TCGA by the presence of more
sample types.  The default type is 'TP' for primary solid tumor.
Codes and their interpretations are available in 
BiocOncoTK::pancan_sampTypeMap.
}
\examples{
if (interactive() && Biobase::testBioCConnection()) {
   billco = Sys.getenv("CGC_BILLING")
   if (nchar(billco)>0) {
     bq = pancan_BQ()
     methSE_BLCA = try(buildPancanSE(bq))
     methSE_BLCA
   }
}
}
