% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_70138}
\alias{cell_70138}
\title{cell_70138: a table with cell-line information from LINCS}
\format{
data.frame
}
\source{
GEO GSE70138 GSE70138_Broad_LINCS_cell_info_2017-04-28.txt.gz
}
\usage{
cell_70138
}
\description{
cell_70138: a table with cell-line information from LINCS
}
\examples{
data(cell_70138)
}
\keyword{datasets}
