% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue.R
\name{clueDemos}
\alias{clueDemos}
\title{generate lists to generate clue API queries}
\usage{
clueDemos()
}
\value{
a list of lists of strings with 'where' and substructure as appropriate
}
\description{
generate lists to generate clue API queries
}
\note{
These are converted to JSON (%20 substituted for embedded blanks.
}
\examples{
clueDemos()
}
