% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{darmGBMcls}
\alias{darmGBMcls}
\title{Data in count_lstpm format from Darmanis 2017 (PMC 5810554) single cell RNA-seq in GBM}
\format{
SummarizedExperiment with HDF Object store back end
}
\source{
http://imlspenticton.uzh.ch/robinson_lab/conquer/data-mae/GSE84465.rds
}
\usage{
darmGBMcls
}
\description{
Data in count_lstpm format from Darmanis 2017 (PMC 5810554) single cell RNA-seq in GBM
}
\note{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5810554/ is the main source article.
}
\examples{
BiocOncoTK::darmGBMcls
}
\keyword{datasets}
