% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbuilder.R
\name{featIDMapper}
\alias{featIDMapper}
\title{define assay-specific feature names in a character vector}
\usage{
featIDMapper()
}
\description{
define assay-specific feature names in a character vector
}
\note{
We may want to use Symbol instead of Entrez when retrieving
expression data.  The value of this function is supplied as a
default for \code{\link{buildPancanSE}}'s featIDMap parameter, 
and alternatives can
be selected by passing similarly named vectors in featIDMap.
}
\examples{
featIDMapper()
}
