% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fireMSI}
\alias{fireMSI}
\title{microsatellite instability data in TCGA, collected from curatedTCGAData}
\format{DataFrame}
\source{
firehose via curatedTCGAData; see metadata(BiocOncoTK::fireMSI)
}
\usage{
fireMSI
}
\description{
microsatellite instability data in TCGA, collected from curatedTCGAData
}
\examples{
str(S4Vectors::metadata(BiocOncoTK::fireMSI))
}
\keyword{datasets}
