% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plate.R
\name{get_plates}
\alias{get_plates}
\title{use curatedTCGAData request to acquire plate codes for samples}
\usage{
get_plates(
  tumcode = "BLCA",
  assay = "RNASeq2GeneNorm",
  samptypes = c("01", "02", "03", "04", "06", "09", "40")
)
}
\arguments{
\item{tumcode}{a TCGA tumor code, usually 3 or for characters}

\item{assay}{a curatedTCGAData assay code, run curatedTCGAData() to
see a message with available options}

\item{samptypes}{a character vector with codes as
defined at \url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/sample-type-codes}}
}
\value{
a data.frame with a row for each TCGA contribution for the selected
tumor type and assay type
}
\description{
use curatedTCGAData request to acquire plate codes for samples
}
\examples{
if (interactive()) {
 plts_blca_rnaseq = get_plates()
 dim(plts_blca_rnaseq)
 head(plts_blca_rnaseq)
}
}
