% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc3utils.R
\name{ggFeatDens}
\alias{ggFeatDens}
\title{create ggplot for density of starts of a GRanges in an interval}
\usage{
ggFeatDens(
  gr,
  mcolvbl,
  chrname = "chr15",
  start = 20450000,
  end = 20730000,
  binwidth.in = 5000,
  basicfilt = function(data) dplyr::filter(data, Consequence ==
    "non_coding_transcript_exon_variant"),
  ylab.in = "feature\\ndensity",
  slstyle = "UCSC"
)
}
\arguments{
\item{gr}{GRanges instance of interest}

\item{mcolvbl}{character(1) mcols(gr) has this variable that will 
be used to specify different groups for computing/colouring the density traces}

\item{chrname}{character(1) chromosome/seqname}

\item{start}{numeric(1) start of interval}

\item{end}{numeric(1) end of interval}

\item{binwidth.in}{numeric(1) for geom_freqpoly binwidth setting}

\item{basicfilt}{a dplyr::filter operation, defaulting to select non-coding variants in mc3 MAF}

\item{ylab.in}{character(1) label for y axis}

\item{slstyle}{character(1) for GenomeInfoDb::seqlevelsStyle}
}
\value{
ggplot instance
}
\description{
create ggplot for density of starts of a GRanges in an interval
}
\examples{
ggFeatDens
}
