% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc3utils.R
\name{ggFeatureSegs}
\alias{ggFeatureSegs}
\title{generate a ggplot of segments of gene-like regions}
\usage{
ggFeatureSegs(chrname = "chr15", start = 20450000, end = 20730000,
  db = EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75, slstyle = "UCSC",
  ylab.in = "ensembl\\nnoncoding")
}
\arguments{
\item{chrname}{character(1) chromosome tag}

\item{start}{numeric(1) start of interval}

\item{end}{numeric(1) end of interval}

\item{db}{EnsDb instance for example}

\item{slstyle}{character(1) tag for seqlevelsStyle}

\item{ylab.in}{character(1) for use as y axis tag}
}
\value{
ggplot instance
}
\description{
generate a ggplot of segments of gene-like regions
}
\note{
Most annotation is turned off with element_blank()
}
\examples{
ggFeatureSegs
}
