% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc3utils.R
\name{ggMutDens}
\alias{ggMutDens}
\title{make a ggplot with density traces of mutations per base pair, for 'most mutated' tumor types in a given interval}
\usage{
ggMutDens(bq, basicfilt = function(data) dplyr::filter(data, Consequence
  == "non_coding_transcript_exon_variant"), chrname = "15",
  start = 20450000, end = 20730000, project_volume = 5,
  maxnrec = 50000, binwidth = 5000, xlab.in = " ")
}
\arguments{
\item{bq}{bigrquery BigQueryConnection instance}

\item{basicfilt}{a dplyr::filter operation, defaulting to select non-coding variants in mc3 MAF}

\item{chrname}{character(1) chromosome token in NCBI seqlevels style}

\item{start}{numeric(1) base coordinate to start}

\item{end}{numeric(1) base coordinate to end}

\item{project_volume}{numeric(1) tumor types will have 
different numbers of contributions; this parameter tells how many
tumor types to represent, counting down from the most frequently represented}

\item{maxnrec}{numeric(1) for as.data.frame}

\item{binwidth}{numeric(1) passed to geom_freqpoly}

\item{xlab.in}{character(1) passed to ggplot2::xlab}
}
\value{
instance of ggplot
}
\description{
make a ggplot with density traces of mutations per base pair, for 'most mutated' tumor types in a given interval
}
\examples{
if (interactive()) {
 if (!requireNamespace("ggplot2")) stop("install ggplot2 to run this function")
 bq = try(pancan_BQ())
 if (!inherits(bq, "try-error")) {
  ggMutDens(bq)
  }
 }
}
