% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_c}
\alias{icd10_c}
\title{helper for interpreting ICD-10 codes}
\format{data.frame}
\source{
ICD-10
}
\usage{
icd10_c
}
\description{
helper for interpreting ICD-10 codes
}
\examples{
BiocOncoTK::icd10_c
}
\keyword{datasets}
