% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k23sig}
\alias{k23sig}
\title{a table of 'significant' MSIsensor-score/expression relationships in TCGA}
\format{
data.frame
}
\usage{
k23sig
}
\description{
a table of 'significant' MSIsensor-score/expression relationships in TCGA
}
\note{
provided to demonstrate bipartite graph construction
}
\examples{
head(k23sig)
}
\keyword{datasets}
