% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kang_DNArepair}
\alias{kang_DNArepair}
\title{list of 151 genes annotated as DNA repair pathway members}
\format{named list}
\source{
\url{https://academic.oup.com/jnci/article/104/9/670/872781#supplementary-data}
}
\usage{
kang_DNArepair
}
\description{
list of 151 genes annotated as DNA repair pathway members
}
\note{
The zipped PDF was read using pdftools::pdf_text and then
manually organized.  All gene symbols present in curatedTCGAData
RNASeq2GeneNorm rownames.  The list elements are
ATM, BER, FA.HR, MMR, NER, NHEJ, OTHER, TLS, RECQ, and XLR.
These denote, respectively, ataxia-telangiectasia-mutated,
base excision repair, Fanconi anemia/homologous recombination,
mismatch repair, nucleotide excision repair,
non-homologous end joining, other, translesion synthesis,
recQ helicase pathway, and cross-link repair.
}
\keyword{datasets}
