% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPatel.R
\name{loadPatel}
\alias{loadPatel}
\title{use BiocFileCache discipline to acquire patelGBMSC SummarizedExperiment}
\usage{

  loadPatel(remotePath = "https://s3.us-east-2.amazonaws.com/biocfound-scrna/patelGBMSC.rds",
  cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{remotePath}{character(1) identifying remote RDS}

\item{cache}{instance of BiocFileCache, defaults to BiocFileCache::BiocFileCache()}
}
\value{
a SummarizedExperiment instance
}
\description{
use BiocFileCache discipline to acquire patelGBMSC SummarizedExperiment
}
\note{
The RDS for the SummarizedExperiment is in an AWS S3 bucket.
This function will check local cache for the data and will download
to cache if not found.  That download is a one-time operation for
any given value of \code{cache}.
}
\examples{
loadPatel
}
