% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{log10pl1}
\alias{log10pl1}
\title{log10(x+p) transformation for use with scales/ggplot2}
\usage{
log10pl1(p = 1)
}
\arguments{
\item{p}{value of shift before taking log10}
}
\value{
an instance of custom trans() for scales package
}
\description{
log10(x+p) transformation for use with scales/ggplot2
}
