% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_tcga_ncit}
\alias{map_tcga_ncit}
\title{a manually constructed table mapping TCGA acronyms to NCIT thesaurus tags}
\format{
data.frame
}
\usage{
map_tcga_ncit
}
\description{
a manually constructed table mapping TCGA acronyms to NCIT thesaurus tags
}
\note{
Constructed using ontoProc::getOncotreeOnto() result.  See the vignette
on Mapping TCGA tumor codes to NCIT for elaborating the mapping to
aggregate tumors into NCIT organ systems.
}
\keyword{datasets}
