% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molpo_3utr}
\alias{molpo_3utr}
\title{representation of 3'UTR MSI events in TCGA from Cortes-Ciriano et al. 2017}
\format{SummarizedExperiment}
\source{
\url{https://www.nature.com/articles/ncomms15180#Sec22}
}
\usage{
molpo_3utr
}
\description{
representation of 3'UTR MSI events in TCGA from Cortes-Ciriano et al. 2017
}
\note{
Supplementary data 6 from publication noted in Source.  See
metadata() component of this SummarizedExperiment for more details.
}
\examples{
molpo_3utr
}
\keyword{datasets}
