% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molpo_CDS}
\alias{molpo_CDS}
\title{representation of MSI events in coding regions TCGA from Cortes-Ciriano et al. 2017}
\format{
SummarizedExperiment
}
\source{
\url{https://www.nature.com/articles/ncomms15180#Sec22}
}
\usage{
molpo_CDS
}
\description{
representation of MSI events in coding regions TCGA from Cortes-Ciriano et al. 2017
}
\note{
Supplementary data 5 from publication noted in Source. See
metadata() component of this SummarizedExperiment for more details.
}
\examples{
molpo_CDS
}
\keyword{datasets}
