% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{molpo_WGS}
\alias{molpo_WGS}
\title{representation of events detected in 708 WGS experiments TCGA from Cortes-Ciriano et al. 2017}
\format{
SummarizedExperiment
}
\source{
\url{https://www.nature.com/articles/ncomms15180#Sec22}
}
\usage{
molpo_WGS
}
\description{
representation of events detected in 708 WGS experiments TCGA from Cortes-Ciriano et al. 2017
}
\note{
Supplementary data 10 from publication noted in Source. See
metadata() component of this SummarizedExperiment for more details.
}
\examples{
molpo_WGS
}
\keyword{datasets}
