% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSIsensor.10k}
\alias{MSIsensor.10k}
\title{MSIsensor microsatellite instability scores for TCGA, collected from Ding et al. Cell 173(2) 2018.}
\format{DataFrame}
\source{
\url{https://www.cell.com/cms/10.1016/j.cell.2018.03.033/attachment/0ac495ba-3578-41cf-8fb1-94487f554bb5/mmc5.xlsx} retrieved 9/17/2018.
}
\usage{
MSIsensor.10k
}
\description{
MSIsensor microsatellite instability scores for TCGA, collected from Ding et al. Cell 173(2) 2018.
}
\examples{
str(BiocOncoTK::dingMSI)
}
\keyword{datasets}
