% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancan.clin.varnames}
\alias{pancan.clin.varnames}
\title{pancan.clin.varnames: a data.frame with a list of variable names for clinical patient data}
\format{
data.frame
}
\source{
pancancer-atlas in BigQuery
}
\usage{
pancan.clin.varnames
}
\description{
pancan.clin.varnames: a data.frame with a list of variable names for clinical patient data
}
\examples{
BiocOncoTK::pancan.clin.varnames[1:5,]
}
\keyword{datasets}
