% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan.R
\name{pancan_app}
\alias{pancan_app}
\title{provide a shiny app to 'glimpse' structure and content of pancan atlas}
\usage{
pancan_app(dataset = "Annotated", nrecs = 5)
}
\arguments{
\item{dataset}{character(1) name of a BigQuery dataset in the pancan-atlas project}

\item{nrecs}{numeric(1) number of records to request (limited through the n= parameter to as.data.table}
}
\value{
currently only as a side effect of starting app
}
\description{
provide a shiny app to 'glimpse' structure and content of pancan atlas
}
\examples{
if (interactive()) pancan_app()
}
