% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan.R
\name{pancan_clinicalTabVarnames}
\alias{pancan_clinicalTabVarnames}
\title{give an interface to tablenames}
\usage{
pancan_clinicalTabVarnames()
}
\value{
interactive datatable from DT
}
\description{
give an interface to tablenames
}
\examples{
if (interactive()) pancan_clinicalTabVarnames()
}
