% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan.R
\name{pancan_longname}
\alias{pancan_longname}
\title{utility to help find long table names}
\usage{
pancan_longname(guess, ...)
}
\arguments{
\item{guess}{a regexp to match the table of interest}

\item{\dots}{passed to \code{\link[base]{agrep}}}
}
\value{
character vector of matches
}
\description{
utility to help find long table names
}
\note{
Note that ignore.case=TRUE is set in the function.
}
\examples{
pancan_longname("rnaseq")
}
