% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancan_sampTypeMap}
\alias{pancan_sampTypeMap}
\title{helper for interpreting pancan-atlas sample type codes}
\format{data.frame}
\source{
ISB BigQuery pancan-atlas project
}
\usage{
pancan_sampTypeMap
}
\description{
helper for interpreting pancan-atlas sample type codes
}
\note{
The sample type codes are not straightforward to interpret.
Primary solid tumor is denoted "TP", and metastatic samples are 
denoted "TM".  This data frame pairs code and natural language terms.
}
\examples{
BiocOncoTK::pancan_sampTypeMap
}
\keyword{datasets}
