% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan.R
\name{pancan_tabulate}
\alias{pancan_tabulate}
\title{tabulate a variable in a table}
\usage{
pancan_tabulate(dataset = "Annotated", tblname, vblname)
}
\arguments{
\item{dataset}{character(1) dataset name}

\item{tblname}{character(1) table name in dataset}

\item{vblname}{character(1) field name in table}
}
\value{
instance of tbl_dbi, constituting summarise result
}
\description{
tabulate a variable in a table
}
\examples{
if (interactive()) pancan_tabulate(tblname=
    "clinical_PANCAN_patient_with_followup", vblname="icd_10")
}
