% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patient_to_tumor_code}
\alias{patient_to_tumor_code}
\title{data.frame mapping from TCGA patient_barcode to TCGA tumor code}
\format{data.frame}
\source{
\url{https://portal.gdc.cancer.gov/exploration?uploadCaseTab=matched}
}
\usage{
patient_to_tumor_code
}
\description{
data.frame mapping from TCGA patient_barcode to TCGA tumor code
}
\note{
Used IDs recorded in MSISensor.10k; one is unmatched at TCGA portal
metadata() component of this SummarizedExperiment for more details.
}
\examples{
head(patient_to_tumor_code)
}
\keyword{datasets}
