% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pert_70138}
\alias{pert_70138}
\title{pert_70138: a table with perturbagen information from LINCS}
\format{
data.frame
}
\source{
GEO GSE70138 GSE70138_Broad_LINCS_pert_info.txt.gz
}
\usage{
pert_70138
}
\description{
pert_70138: a table with perturbagen information from LINCS
}
\examples{
data(pert_70138)
}
\keyword{datasets}
