% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue.R
\name{pertClasses}
\alias{pertClasses}
\title{enumerate perturbagen classes}
\usage{
pertClasses(key = Sys.getenv("CLUE_KEY"))
}
\arguments{
\item{key}{character(1) API key provided by clue.io}
}
\value{
a character vector
}
\description{
enumerate perturbagen classes
}
\examples{
if (nchar(Sys.getenv("CLUE_KEY"))>0) {
  pc = pertClasses()
  head(vapply(pc, "[[", character(1), 1))
}
}
