% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue.R
\name{query_clue}
\alias{query_clue}
\title{run the api.clue.io API to acquire information on LINCS experiments}
\usage{
query_clue(service = "profiles", filter = list(where = (list(pert_iname =
  "sirolimus", cell_id = "MCF7", assay = "L1000"))),
  key = Sys.getenv("CLUE_KEY"))
}
\arguments{
\item{service}{a character(1) service name}

\item{filter}{a list to be converted to JSON for submission as a GET request}

\item{key}{character(1) API key provided by clue.io}
}
\description{
run the api.clue.io API to acquire information on LINCS experiments
}
\examples{
if (nchar(Sys.getenv("CLUE_KEY"))>0) {
demos = clueDemos()
nd = length(demos)
chk = lapply(seq_len(nd), function(x) query_clue( service=names(demos)[x],
              filter=demos[[x]]) )
names(chk) = names(demos)
sapply(chk,length)
}
}
