% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbuilder.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{filter}
\alias{tbl}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{tbl}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

