% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbuilder.R
\name{replaceRownames}
\alias{replaceRownames}
\title{map rownames of an SE to another vocabulary}
\usage{
replaceRownames(se, sourceVocab = "ENTREZID", targetVocab = "SYMBOL")
}
\arguments{
\item{se}{SummarizedExperiment instance}

\item{sourceVocab}{character(1) must be a keytype of org.Hs.eg.db, defaults to 'ENTREZID'}

\item{targetVocab}{character(1) must be a column of org.Hs.eg.db}
}
\description{
map rownames of an SE to another vocabulary
}
