% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_msi}
\alias{small_msi}
\title{filtered MSI data for demonstrating exploratory app}
\format{
DataFrame
}
\source{
MSI values from dingMSI, expression from curatedTCGAData for three genes, two tumors
}
\usage{
small_msi
}
\description{
filtered MSI data for demonstrating exploratory app
}
\examples{
head(BiocOncoTK::small_msi)
}
\keyword{datasets}
