% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprintISB.R
\name{TcgaNIndWithAnyMut}
\alias{TcgaNIndWithAnyMut}
\title{Give count of individuals with a mutation recorded for selected tumor}
\usage{
TcgaNIndWithAnyMut(
  tumor = "BRCA",
  limit = NULL,
  db = "isb-cgc:tcga_201607_beta",
  project
)
}
\arguments{
\item{tumor}{character(1) defaults to 'BRCA'}

\item{limit}{numeric(1) defaults to NULL, appended as limit to number of records returned if non-null}

\item{db}{character(1) BigQuery database name}

\item{project}{character(1) project code}
}
\value{
numeric(1)
}
\description{
Give count of individuals with a mutation recorded for selected tumor
}
\examples{
if (interactive()) TcgaNIndWithAnyMut(project="cgc-05-0009")
}
