% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan.R
\name{tumNorSet}
\alias{tumNorSet}
\title{create list with SEs for tumor and normal for a tumor/assay pairing}
\usage{
tumNorSet(bq, code = "PRAD",
  assayDataTableName = pancan_longname("rnaseq"),
  assayValueFieldName = "normalized_count",
  assayFeatureName = "Entrez")
}
\arguments{
\item{bq}{a BigQuery connection}

\item{code}{character(1) a TCGA tumor code, defaults to "PRAD" for prostate tumor}

\item{assayDataTableName}{character(1) name of table in BigQuery}

\item{assayValueFieldName}{character(1) field from which assay quantifications are retrieved}

\item{assayFeatureName}{character(1) field from which assay feature names are retrieved}
}
\description{
create list with SEs for tumor and normal for a tumor/assay pairing
}
\examples{
if (interactive()) {
 bqcon = try(pancan_BQ())
 if (!inherits(bqcon, "try-error")) {
   tn = tumNorSet(bqcon)
   tn
 }
}
}
