% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_ccleNRAS}
\alias{load_ccleNRAS}
\alias{load_NRAS_AHR}
\alias{load_nrasdf}
\title{utilities for mock data (not involving internet access for vignette)}
\usage{
load_ccleNRAS()

load_NRAS_AHR()

load_nrasdf()
}
\value{
a list of DRProfSet instances

a data.frame with fields 'Cell_line_primary_name', 'RMA_normalized_expression', 'HGNC_gene_symbol'

a data.frame
}
\description{
utilities for mock data (not involving internet access for vignette)
}
\note{
These functions are provided only for avoiding reliance on internet connectivity for document production.
}
\examples{
load_ccleNRAS()
dim(load_nrasdf())
}
