% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bindMSI}
\alias{bindMSI}
\title{bind MSI data to a SummarizedExperiment}
\usage{
bindMSI(se)
}
\arguments{
\item{se}{SummarizedExperiment instance}
}
\value{
SummarizedExperiment instance with expanded colData,
samples limited to those with microsatellite instability values
}
\description{
bind MSI data to a SummarizedExperiment
}
\examples{
bindMSI
}
