% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEbuilder.R
\name{buildPancanSE}
\alias{buildPancanSE}
\title{helper for SummarizedExperiment construction from pancan}
\usage{
buildPancanSE(bq, acronym = "BLCA", assay = "meth450k",
  sampType = "TP", subjectIDName = "ParticipantBarcode")
}
\arguments{
\item{bq}{instance of BigQueryConnection for pancancer-atlas.Annotated Dataset}

\item{acronym}{character(1) 'cohort' label, e.g., 'BLCA'}

\item{assay}{character(1) element from names(BiocOncoTK::annotTabs), e.g., 'meth450k'}

\item{sampType}{character(1) element from 
BiocOncoTK::pancan_sampTypeMap$"SampleTypeLetterCode", 
e.g., 'TP' for Primary solid Tumor samples,
or 'TB' for peripheral blood sample from primary blood derived cancer}

\item{subjectIDName}{character(1) field name for subject identifier}
}
\value{
SummarizedExperiment
}
\description{
helper for SummarizedExperiment construction from pancan
}
\note{
Note that pancancer-atlas is distinguished from TCGA by the presence of more
sample types.  The default type is 'TP' for primary solid tumor.
Codes and their interpretations are available in BiocOncoTK::pancan_sampTypeMap.
}
\examples{
if (interactive() && Biobase::testBioCConnection()) {
   billco = Sys.getenv("CGC_BILLING")
   if (nchar(billco)>0) {
     bq = pancan_BQ()
     methSE_BLCA = try(buildPancanSE(bq))
     methSE_BLCA
   }
}
}
