% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clue.R
\name{clueServiceNames}
\alias{clueServiceNames}
\title{Provide names of some clue.io services for which examples are available in this package.}
\usage{
clueServiceNames()
}
\value{
a character vector of service names
}
\description{
Provide names of some clue.io services for which examples are available in this package.
}
\note{
See \url{https://clue.io/api}.
}
\examples{
clueServiceNames()
}
