% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoprintISB.R
\name{oncoPrintISB}
\alias{oncoPrintISB}
\title{interactive interface to ComplexHeatmap oncoPrint with inputs from ISB Cancer Genomics Cloud BigQuery back end}
\usage{
oncoPrintISB(bq)
}
\arguments{
\item{bq}{an instance of \code{\link[bigrquery]{BigQueryConnection-class}} authenticated for ISB Cancer Genomics Cloud access}
}
\value{
only used for side effect of running shiny app
}
\description{
interactive interface to ComplexHeatmap oncoPrint with inputs from ISB Cancer Genomics Cloud BigQuery back end
}
\note{
This function will start a shiny app and will generate queries to
Google BigQuery tables representing TCGA.
}
\examples{
if (interactive()) {
 bcode = Sys.getenv("CGC_BILLING")
 if (nchar(bcode)>0) {
  con <- DBI::dbConnect(bigrquery::dbi_driver(), project = "isb-cgc", 
       dataset = "tcga_201607_beta", billing = bcode)
  oncoPrintISB(con)
 }
}
}
